Conflicts:
	common/src/main/java/com/scholarscore/models/Section.java
	common/src/test/java/com/scholarscore/models/GradeFormulaTest.java
	warehouse/src/main/java/com/scholarscore/api/controller/StudentAssignmentController.java
	warehouse/src/main/java/com/scholarscore/api/manager/StudentAssignmentManager.java
	warehouse/src/main/java/com/scholarscore/api/persistence/StudentAssignmentPersistence.java
	warehouse/src/main/java/com/scholarscore/api/persistence/goalCalculators/AssignmentGoalCalc.java
	warehouse/src/main/java/com/scholarscore/api/persistence/goalCalculators/AttendanceGoalCalc.java
	warehouse/src/main/java/com/scholarscore/api/persistence/mysql/jdbc/StudentAssignmentJdbc.java
	warehouse/src/test/java/com/scholarscore/api/util/SchoolDataFactory.java

+++ warehouse/src/main/java/com/scholarscore/api/persistence/goalCalculators/AttendanceGoalCalc.java
@@ -25,7 +25,7 @@ public class AttendanceGoalCalc implements GoalCalc<CalculatableAttendance> {
                     dayAttendance.getAssignment().getDueDate().before(goal.getStartDate())) {
                 continue;
             }
-            Long points = dayAttendance.getAwardedPoints();
+            Double points = dayAttendance.getAwardedPoints();
             if (null == points) {
 
             } else if (points == 0) {
